requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.26
langsmith>=0.1.0
openai>=1.6.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
twilio>=8.10.0
python-dateutil>=2.8.0
schedule>=1.2.0
typing-extensions>=4.9.0
